update autoexec callback && invoke

+++ src/main/java/codedriver/module/autoexec/api/job/AutoexecJobDeleteApi.java
@@ -12,7 +12,7 @@ import codedriver.framework.common.constvalue.ApiParamType;
 import codedriver.framework.restful.annotation.*;
 import codedriver.framework.restful.constvalue.OperationTypeEnum;
 import codedriver.framework.restful.core.privateapi.PrivateApiComponentBase;
-import codedriver.module.autoexec.dao.mapper.AutoexecJobMapper;
+import codedriver.framework.autoexec.dao.mapper.AutoexecJobMapper;
 import codedriver.module.autoexec.service.AutoexecJobActionService;
 import codedriver.module.autoexec.service.AutoexecJobService;
 import com.alibaba.fastjson.JSONObject;
@@ -60,8 +60,10 @@ public class AutoexecJobDeleteApi extends PrivateApiComponentBase {
     public Object myDoService(JSONObject jsonObj) throws Exception {
         Long jobId = jsonObj.getLong("jobId");
         AutoexecJobVo jobVo = autoexecJobMapper.getJobLockByJobId(jobId);
-        autoexecJobActionService.executeAuthCheck(jobVo);
-        jobService.deleteJob(jobId);
+        if(jobVo != null) {
+            autoexecJobActionService.executeAuthCheck(jobVo);
+            jobService.deleteJob(jobId);
+        }
         return null;
     }
